consolidate chat input components and clean up console logs
